requirements.txt
absl-py==2.1.0
academictorrents==2.3.3
accelerate==1.6.0
addict==2.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
albucore==0.0.23
albumentations==2.0.5
alembic==1.15.1
annotated-types==0.7.0
ansi2html==1.8.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.1.0
audioread==3.0.1
banal==1.0.6
beautifulsoup4==4.13.4
bencode.py==2.0.0
biopython==1.85
bitstring==3.1.5
blobfile==3.0.0
braceexpand==0.1.7
brokenaxes==0.6.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
DataProperty==1.1.0
dataset==1.6.2
datasets==3.3.2
decorator==5.2.1
diffusers==0.32.2
dill==0.3.8
distro==1.9.0
docstring_parser==0.16
easydict==1.13
einops==0.8.1
et_xmlfile==2.0.0
evaluate==0.4.3
exceptiongroup==1.2.2
executing==2.2.0
ExifRead==2.3.2
fastapi==0.115.14
fastmcp==2.4.0
filelock==3.17.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
future==0.16.0
greenlet==3.1.1
grpcio==1.70.0
h11==0.14.0
hf-xet==1.1.8
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.34.4
idna==3.10
imageio==2.37.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
iniconfig==2.1.0
ipython==8.36.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonargparse==4.37.0
jsonlines==4.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.10.2.post1
lightning==2.2.5
lightning-utilities==0.13.1
llvmlite==0.44.0
lm_eval==0.4.4
loguru==0.7.3
lpips==0.1.4
lxml==5.4.0
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mbstrdecoder==1.1.4
mcp==1.9.0
mdurl==0.1.2
MedPy==0.5.2
ml_collections==1.1.0
modelscope==1.25.0
monai==1.4.0
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.2.1
nibabel==5.3.2
ninja==1.11.1.4
nltk==3.9.1
numba==0.61.0
numexpr==2.10.2
numpy==1.25.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==13.580.65
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.1.105
nvitop==1.5.3
omegaconf==2.3.0
openai==1.72.0
openapi-pydantic==0.5.1
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
packaging==24.2
pandas==2.3.0
parso==0.8.4
pathvalidate==3.2.3
patsy==1.0.1
peft==0.15.2
pexpect==4.9.0
pidng==3.4.7
pillow==11.1.0
piq==0.8.0
platformdirs==4.3.6
pluggy==1.6.0
pooch==1.8.2
portalocker==3.1.1
prompt_toolkit==3.0.51
propcache==0.3.0
protobuf==6.30.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-spy==0.3.14
pyarrow==19.0.1
pybind11==2.13.6
pycparser==2.22
pycryptodomex==3.22.0
pydantic==2.10.6
pydantic-settings==2.9.1
pydantic_core==2.27.2
Pygments==2.15.1
PyMuPDF==1.23.26
PyMuPDFb==1.23.22
pyparsing==3.2.1
PyPubSub==3.3.0
PySide6-Essentials==6.5.2
pytablewriter==1.2.1
pytest==7.4.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytorch-ignite==0.5.1
pytorch-lightning==2.5.0.post0
pytorch-msssim==1.0.0
pytz==2025.1
PyYAML==6.0.2
ray==2.49.1
referencing==0.36.2
regex==2023.6.3
requests==2.32.3
rich==14.0.0
rouge_score==0.1.2
rpds-py==0.27.1
sacrebleu==2.5.1
safetensors==0.5.3
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.2
shellingham==1.5.4
shiboken6==6.5.2
SimpleITK==2.4.1
simplejson==3.20.1
simsimd==6.2.1
six==1.17.0
sniffio==1.3.1
some-package==0.1
sortedcontainers==2.4.0
soundfile==0.13.1
soupsieve==2.7
soxr==0.5.0.post1
SQLAlchemy==1.4.54
sqlitedict==2.1.0
sse-starlette==2.3.5
stack-data==0.6.3
starlette==0.46.2
statsmodels==0.14.4
stringzilla==3.12.2
sympy==1.13.3
tabledata==1.3.4
tabulate==0.9.0
tcolorpy==0.1.7
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==3.1.0
threadpoolctl==3.5.0
tifffile==2025.2.18
tiktoken==0.4.0
timm==1.0.15
tokenizers==0.21.4
tomli==2.2.1
torch==2.4.1
torchinfo==1.8.0
torchmetrics==1.6.2
torchvision==0.19.1
tqdm==4.67.1
tqdm-multiprocess==0.0.11
traitlets==5.14.3
transformers==4.50.0
triton==3.0.0
typepy==1.3.4
typer==0.15.4
typeshed_client==2.7.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.1
Unidecode==1.3.8
urllib3==2.3.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.1.0
wcwidth==0.2.13
webdataset==0.2.111
websockets==15.0.1
Werkzeug==3.1.3
word2number==1.1
xxhash==3.5.0
yacs==0.1.8
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
